Update User Model to Fix Parameter Signature (#24)
Replaces the updated time variable type with Long from previously assigned Int

Since this field will be most likely an epoch time stamp

+++ data/src/main/java/co/anitrend/data/model/response/core/user/User.kt
@@ -47,5 +47,5 @@ data class User(
     override val siteUrl: String?,
     override val donatorTier: Int?,
     override val moderatorStatus: String?,
-    override val updatedAt: Int?
+    override val updatedAt: Long?
 ) : IUser
