+++ data/src/main/java/co/anitrend/data/model/response/core/user/contract/IUser.kt
@@ -48,5 +48,5 @@ interface IUser : IEntity {
     val siteUrl: String?
     val donatorTier: Int?
     val moderatorStatus: String?
-    val updatedAt: Int?
+    val updatedAt: Long?
 }
